Exploring top down syntax parsing
Start of section interior and others using topdown functional
decomposition. Focusing on recognition first, (then error
complaints,) then building a parse tree.

Removes Name from lexeme list (as it's more of a syntaxeme).
Adds syntaxeme, though I'm not using it yet. Toss had an
invalid comment as an initial line (which this will handle,
but not as a first attempt).