Parser state diagram distinguishes field_value from field_any
Field_any represents either the "empty" field, a copied field,
or a temporary phase that switches to field_value when it
encounters more text on the same line. The previous conception
didn't distinguish enough for cases where the field started with
some text, but was before an (errant) map, which isn't valid.