Adds a candidate parse() to classify line by line.
I'll defer affirming that the lines should follow one another to
another phase. Needs name, escaped name, section, and
the interior of multiline. Adds the empty syntaxeme, in case
I want to honor other eno librarys' ability to guess location
in the original document (or to emit similar to original).