Renames endl continue op to empty; groups parse methods close
In anticipation of break continuation operator becoming an empty
continuation operator, this renames the corresponding lexeme.
As I keep forgetting which is which while writing tests, I included
javadoc to that effect.

Moves the deprecated methods for recognizing in Parser down
together so I can browse the active parse methods quickly.