Moves relevant classes to parse package
The base package had become cluttered, or would when I
add two or three more EnoElement classes. This extracts
the serialization related classes to a parse package.

Semantologist.field() uses EnoType to track which field
it becomes, rather than the Syntaxeme. Unfortunately,
section in syntaxeme and enotype collide, so I needed
to qualify it.