Clone comments, initial values; semantic field() adds value
Adds placeholders for field of list and field of set. Field
allows cloning from another field. Corrects the multiline
constructor that used the empty field type, instead of multiline.
Multiline and Value use the canon method name for getting
the value.

Semantologist.field() partially handles appending a value to
a Value or creating a Value from Empty. Still needs to add
value to latest entry or item of set/list, correspondingly.

module-info notes that it hides the parse package.