Completes field() draft; tracks more aspects
Wraps fieldset section of field() . EnoElement records the
template name escapes and the line. Field clone involves
more (sub)fields. FieldSet stores and exposes entries. Spec
says that the name of list items is the name of the field they
are inside of. Journal complains that the canon eno locale
project changed to invalidate note_enojes.