Began missing element responses to requesting section children
In attempting to get a field from the document - section returned
by Semantologist, I noticed that the canon api returns either
null, an exception, or a fake element. This seems
straightforward enough to accommodate.

Each fake version of each element will be largely the same. In the
future, I may ensure that the complaint methods for Bomb are
default interface methods. The classes of the missing package are
mostly to return from Section.getSpecificChild().