Adds null and bounds checks so hand logging doesn't die
Adds else clauses as well so I can see when the
returned words are null or I am beyond the boundary.
This currently isn't parsing the sub test on line 109
of ShouldSemantics (a field with a continued value),
indicating a problem with field() noticing that the
next element is a child. Perhaps I need to peek at the
current line or something, like I needed to in
buildDocument() to advance.