Adds hetero style doc test; advance for field; incomplete
Incomplete; currently didn't add expected children to
a section. I may have been too aggressive.

ShouldSemantics now has a subtest for a document of
multiple levels of elements, including blank lines and
an associated comment. Uses infix increment for line to
avoid hand calculating the line number. Uses DocGen for
the newest test.

DocGen is a wrapper around StringBuilder that allows
creating an eno document, without constraining itself to
only emitting valid documents. While I intend to use it
as a method chaining means of building strings for tests,
it offers a means of creating a document without obliging
the use of actual eno elements. This has relevance
because I'm not (yet?) exposing Lexeme and its characters
for writing documents.

Corrects comment cloning in EnoElement, as I'd copied to
the caller's rather than the destination list of
comments.

Semantologist.section() advances the line before trying to
add children of appropriate types. It also checks for the
preceeding comment on that line, rather than past it,
which may advance the cursor. I've not yet proven it's a
problem.