Associates comments with subfield; adds package-info
Grammarian.buildDocument() looks at the current line
for the next type. Reorganizes some if blocks in
peekAtNextLineType() for clarity.

DocGen explicitly creates a list, in case the caller
wants to change it. I've been disabused of the need for
attempting that, but it's roughly the same.

Adds better tests to ShouldSemantics for associated comments.