Multiline is a Value subclass
To avoid replicating aspects that are the same, in a
situation where I'm not enforcing the difference, this
makes Multiline a child of Value, rather than Field.
DocGen is able to create multiline of desired boundary
and no body. Grammarian follows the spec of treating
an empty body for multiline as null.