Actually honors template; permits first space in value
Grammarian adds the prepared Dependency to transitive x.
FieldList.getOnlyValues() uses the combined items,
rather than just the local items. Minor improvement
to FL.setTemplate() when provided an EnoElement, rather
than a cast FieldList (as when Grammarian is resolving
forward references).

Parser previously trimmed the first space of a phrase
value (hmm all my tests are single word). This checks
whether the phrase is a single word and trims that or
only trims the edges of the first word and the rest.
Exposes trimming for testing.

Removes package info from the test packages, as they
didn't add much and confused my older eclipse.